Add turkish translation to tr/profile.json
Hello,
There's a code that we would need to change, because of the different structure of phrases in Turkish.
From line 47 to 50= there are the words "can"/"cannot", but there is no such words in Turkish. 
Is it possible to delete those lines (47,48,49), where the sentences in divided in different pieces. And instead, put the whole sentences, like in line 50:
- line 50 is already translated for 'Kiron CAN contact me..."
- new line to create for 'Kiron CANNOT contact me". Turkish version: Kiron, WhatsApp Messenger aracılığıyla benimle iletişim kuramaz.
Thank you